bug fix:goblin

+++ empery_center/src/cluster/map.cpp
@@ -670,7 +670,7 @@ _wounded_done:
 		try {
 			Poseidon::enqueue_async_job([=]() mutable {
 				PROFILE_ME;
-
+				
 				if(attacked_object_type_id == MapObjectTypeIds::ID_WORLD_ACTIVITY_BOSS){
 					return;
 				}
@@ -679,13 +679,14 @@ _wounded_done:
 				if(!monster_type_data){
 					return;
 				}
-
+				
 				bool goblin_award = false;
 				static constexpr auto GOBLIN_WEAPON_ID = MapObjectWeaponId(2605001);
 				if(monster_type_data->map_object_weapon_id == GOBLIN_WEAPON_ID){
 					goblin_award = true;
 				}
-
+				
+				
 				const auto item_box = ItemBoxMap::get(attacking_account_uuid);
 				if(!item_box){
 					LOG_EMPERY_CENTER_DEBUG("Failed to load item box: attacking_account_uuid = ", attacking_account_uuid);
@@ -833,7 +834,7 @@ _wounded_done:
 					const auto hp_damaged_now = checked_sub(hp_total,hp_remaining);
 					const auto hp_damaged_last = checked_sub(hp_damaged_now,hp_damaged);
 					const auto interval  = hp_total*Data::Global::as_double(Data::Global::SLOT_GOBLIN_DROP_AWARD_HP_PERCENT);
-					auto reward_count = static_cast<int>(hp_damaged_now/interval - hp_damaged_last/interval);
+					auto reward_count = static_cast<int>(hp_damaged_now/interval) - static_cast<int>(hp_damaged_last/interval);
 					bool real_award = false;
 					if(hp_damaged_now == hp_total){
 						//最后死亡走普通怪物掉落
@@ -1188,13 +1189,13 @@ _wounded_done:
 			LOG_EMPERY_CENTER_ERROR("std::exception thrown: what = ", e.what());
 		}
 	}
-
+	
 	//世界活动刷怪
 	if(attacking_account_uuid && (soldiers_remaining == 0)){
 		Poseidon::enqueue_async_job([=]{
 			try {
 				PROFILE_ME;
-
+	
 				const auto monster_type_data = Data::MapObjectTypeMonster::get(attacked_object_type_id);
 				if(!monster_type_data){
 					return;
